Move windows code to win folder in brave_vpn component
