+++ components/brave_vpn/brave_vpn_service.cc
@@ -13,9 +13,9 @@
 #include "base/json/json_writer.h"
 #include "base/strings/utf_string_conversions.h"
 #include "base/time/time.h"
-#include "brave/components/brave_vpn/brave_vpn_constants.h"
 #include "brave/components/brave_vpn/brave_vpn_service_helper.h"
 #include "brave/components/brave_vpn/brave_vpn_utils.h"
+#include "brave/components/brave_vpn/common/brave_vpn_constants.h"
 #include "brave/components/brave_vpn/pref_names.h"
 #include "brave/components/brave_vpn/vpn_response_parser.h"
 #include "brave/components/p3a_utils/feature_usage.h"
