+++ components/brave_vpn/common/BUILD.gn
@@ -1,0 +1,13 @@
+# Copyright (c) 2022 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import("//brave/components/brave_vpn/buildflags/buildflags.gni")
+
+assert(enable_brave_vpn)
+
+source_set("common") {
+  sources = [ "brave_vpn_constants.h" ]
+  deps = [ "//build:buildflag_header_h" ]
+}
