+++ components/brave_vpn/win/BUILD.gn
@@ -1,0 +1,49 @@
+# Copyright (c) 2022 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import("//brave/components/brave_vpn/buildflags/buildflags.gni")
+
+assert(enable_brave_vpn)
+assert(is_win)
+
+source_set("utils") {
+  sources = [
+    "utils.cc",
+    "utils.h",
+  ]
+  libs = [ "rasapi32.lib" ]
+  deps = [
+    "//base",
+    "//brave/components/brave_vpn/common",
+  ]
+}
+
+source_set("win") {
+  visibility = [ "//brave/components/brave_vpn/*" ]
+
+  sources = [
+    "brave_vpn_os_connection_api_win.cc",
+    "brave_vpn_os_connection_api_win.h",
+  ]
+
+  deps = [
+    ":utils",
+    "//base",
+    "//brave/components/brave_vpn/common",
+    "//brave/components/brave_vpn/mojom",
+    "//net",
+    "//services/network/public/cpp",
+  ]
+}
+
+executable("vpntool") {
+  sources = [ "winvpntool.cc" ]
+  libs = [ "rasapi32.lib" ]
+  deps = [
+    ":win",
+    "//base",
+    "//brave/components/brave_vpn/win:utils",
+  ]
+}
