+++ components/brave_vpn/win/brave_vpn_os_connection_api_win.h
