+++ components/brave_vpn/win/utils.cc
@@ -1,9 +1,9 @@
 /* Copyright (c) 2021 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_vpn/utils_win.h"
+#include "brave/components/brave_vpn/win/utils.h"
 
 #include <windows.h>
 
@@ -15,6 +15,7 @@
 #include "base/logging.h"
 #include "base/path_service.h"
 #include "base/strings/stringprintf.h"
+#include "brave/components/brave_vpn/common/brave_vpn_constants.h"
 
 #define DEFAULT_PHONE_BOOK NULL
