+++ components/brave_vpn/win/utils.h
